Fall back to non-parallelized tests The combination of pytest xfail and xdist are causing our helper cleanup function to be called before the rest of the suite has finished executing. This results in missing docker images (since cleanup has removed them), triggering failures elsewhere in the test suite when xfail is encountered under certain conditions.
I've spent some time trying to debug it, but have been unable to arrive at a satisfactory solution. Fall back to single process for xdist in order to work around the problem. I'll be opening a GH issue to track this, as we will certainly want the tests to be parallelizable as the suite grows.