Clarify that X.509-SVID Signing Certificate may be an intermediate There has been some confusion over whether or not SPIFFE IDs must be set from the root down. The intention of the current text is that only the most immediate signing certificate, which may be an intermediate, must have a SPIFFE ID. The issuer of the signing certificate is not required to be an SVID.
We should additionally revisit the MUST which mandates that signing certificates themselves be an SVID. I think that this makes sense logically, though the MUST was put in place at a time when the text described its use during the validation process - this is no longer the case. This matter probably warrants further discussion in SIG-Spec.