edit SPIFFE license info so GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the contents of the LICENSE file with the full
text of the Apache 2.0 license. This allows Licensee to successfully
identify the license type (as Apache 2.0).