Introduce the SPIFFE Trust Domain and Bundle specification This commit introduces the SPIFFE Trust Domain and Bundle specification.
The specification describes the semantics of SPIFFE trust domains, how
their authorities are represented, and the mechanism by which they may
be coupled together (i.e. "federation"). It has been developed by
SPIFFE's SIG-Spec community with the supervision of the SPIFFE
maintainers, the members of which are documented in `CODEOWNERS.md`.

This change includes amendments to the X509-SVID and JWT-SVID
specifications. The mechanics of those specifications remain unchanged,
however amendments were necessary to define the manner in which their
authorities are published to and consumed from a SPIFFE bundle.

Special thank you to @justinburke, @spikecurtis, and @mattmoyer for
their efforts in shaping this document.

Signed-off-by: Evan Gilman <evan@scytale.io>