Clarify JWT-SVID serialization requirements @prabath has pointed out that our requirement for JWT-SVIDs to be serialized using JWS Compact Serialization is redundant because the JWT RFC itself does not permit JWS JSON Serialization.
This commit attempts to clarify this point by amending section 5.1 with
text that states that the JWT RFC mandates this already. @prabath has
suggested that section 5.1 be struck in its entirety, however due
diligence by SIG-Spec has shown that there are a handful of popular
libraries that still permit this serialiazation format, thus we thought
it prudent for the comment to remain.

Some minor redundancy has been left in place to aid readers who do not
read the document in its entirety.

Signed-off-by: Evan Gilman <evan@scytale.io>