Update Workload API spec with JWT-SVID Profile
This PR adds the JWT-SVID Profile to the Workload API specification.

Most of the JWT-SVID text came from a draft proposal created by
@evan2645. There have been some small tweaks, mostly adapting the
language based on the recent addition of FetchX509Bundles, which closely
mirrors the FetchJWTBundles RPC.

I also took the opportunity to create a canonical, complete definition
of the Workload API gRPC Service definition. Consequently some of the
service definition text in the spec has been adjusted.

Signed-off-by: Andrew Harding <aharding@vmware.com>