Add SPIFFE Steering Committee Charter The SSC, formerly known as the TSC, has recently revamped their charter to better reflect the needs of the projects. The new charter was developed over a series of public calls, in cooperation with SPIFFE project maintainers. This commit ports the working version out of Google Docs and into Markdown. The original version can be found here: https://docs.google.com/document/d/1OvNBn9s3WBegbgOPUu-YMUzbgzk-CjSR2Gk4-XPgvfk
This commit also updates the CODEOWNERS file to better reflect reality.
A new `ssc` GitHub group will be kept up-to-date, and some files in this
repository are managed by them (vs SPIFFE maintainers).

Finally, the project governance document has been updated to reflect the
new charter and naming.

Please note that the governance document is in need of further update,
however that is out of scope for this change.

Signed-off-by: Evan Gilman <egilman@vmware.com>