Reconcile contradiction between ValidateJWTSVID gRPC definition and spec
The comments in the gRPC definition of ValidateJWTSVID declare the
ValidateJWTSVIDResponse.claims field to be optional, but the
specification text indicates that all fields in ValidateJWTSVIDResponse
are required. These two statements contradict one another.

Updating the text in the specification to match the comment in the gRPC
definition of ValidateJWTSVID using similar language to descriptions of
other Workload API request/response message types.

Signed-off-by: Ryan Turner <turner@uber.com>