Add SPIFFE maturity phases
This commit adds a document that describes the maturity phases of SPIFFE
projects. As SPIFFE grows, and so does the number of (sub)projects,
clearly communicating to our users the maturity level of the software in
question is very important in maintaining user trust.

As discussed in the June SSC call.

Signed-off-by: Evan Gilman <egilman@vmware.com>