Add security considerations text around "claims"
This commit updates text in the SPIFFE ID specification to address the
common question of supporting "claims" in SVIDs. The text attempts to
explain some of the gotchas associated with "claim" use (which we refer
to here as "assertions"), and provides guidelines by which readers may
assess the relative safety of any given assertion.

This text has been authored by SIG-Spec and was ported from the Google
Doc link below. It supersedes #185.

https://docs.google.com/document/d/1gtcQ73Cc24XA6uBgeMCS9IOcUEvYpbj8iTiUiuMFtvE

Signed-off-by: Evan Gilman <egilman@vmware.com>